+++ core/src/main/java/org/dotwebstack/framework/config/FileConfigurationBackend.java
@@ -17,9 +17,9 @@ import javax.annotation.PostConstruct;
 import lombok.NonNull;
 import org.apache.commons.io.FilenameUtils;
 import org.dotwebstack.framework.EnvironmentAwareResource;
-import org.dotwebstack.framework.validate.ShaclValidationException;
-import org.dotwebstack.framework.validate.ShaclValidator;
-import org.dotwebstack.framework.validate.Validator;
+import org.dotwebstack.framework.validation.ShaclValidationException;
+import org.dotwebstack.framework.validation.ShaclValidator;
+import org.dotwebstack.framework.validation.Validator;
 import org.eclipse.rdf4j.RDF4JException;
 import org.eclipse.rdf4j.repository.RepositoryConnection;
 import org.eclipse.rdf4j.repository.RepositoryException;
