+++ core/src/test/java/org/dotwebstack/framework/config/FileConfigurationBackendTest.java
@@ -24,7 +24,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.List;
 import java.util.stream.Collectors;
-import org.dotwebstack.framework.validate.ShaclValidator;
+import org.dotwebstack.framework.validation.ShaclValidator;
 import org.eclipse.rdf4j.repository.RepositoryException;
 import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.eclipse.rdf4j.repository.sail.SailRepositoryConnection;
