+++ integration-test/src/test/java/org/dotwebstack/framework/FileConfigurationBackendPrefixesIntegrationTest.java
@@ -2,7 +2,7 @@ package org.dotwebstack.framework;
 
 import org.dotwebstack.framework.config.ConfigurationException;
 import org.dotwebstack.framework.config.FileConfigurationBackend;
-import org.dotwebstack.framework.validate.ShaclValidator;
+import org.dotwebstack.framework.validation.ShaclValidator;
 import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.eclipse.rdf4j.sail.memory.MemoryStore;
 import org.junit.Before;
